ajaxSpider: use the URI of the selected node for the spider's seed
Change to use the full URI of the selected "Sites" tree node, for the
spider's seed, instead of just up to the path component. Not only fixes
the issue with missing GET parameters but it also normalises the
behaviour with the rest of scans (e.g. "normal" spider and active scan)
Fix #1674 - Ajax spider not considering get parameters