Prepare add-ons for upcoming removal of JBroFuzz libraries from core
With the removal of JBroFuzz libraries from core the add-ons must bundle
the removed libraries, in this case it's enough to just include
jbrofuzz-2.5.1.jar (the jar was obtained from Maven Central Repository).
The affected add-ons, Zest and Fuzz, will now include that libary and
also the file with the fuzzers (whose path is passed to JBroFuzz class).
The JBroFuzz library is removed in favour of jbrofuzz-2.5.2.jar from the
"lib" directory. This change forces the addition of
sqlite-jdbc-3.7.15-M1.jar (copied from core) to lib directory, as it's
a compilation dependency of an add-on and was being added to classpath
by JBroFuzz library.