ascanrules: fix memory leak in PersistentXSSUtils
Change to write the HTTP messages to database, as temporary scan
messages, to avoid having the messages in memory longer than needed.
Change to cache (with SoftReference) the URIs and parameters used for
the sink messages, those are used often and might be new strings which
would increase the memory usage with duplicated Strings.
Change to clear the cached data and messages IDs when a new session is
about to be created (the messages would no longer be available moreover
temporary messages should be automatically deleted when session/database
is closed).
Update changes in ZapAddOn.xml file.
Fix zaproxy/zaproxy#1974 - Memory leak with "Active scanner rules"
add-on