ascanrulesBeta: remove XXE API when the add-on is uninstalled
Change ExtensionAscanRulesBeta to take care to install and uninstall the
API used by XXEPlugin. The API is (created and) installed when required
by the XXEPlugin, and removed when the add-on is uninstalled.
Change XXEPlugin to no longer create and install the API but obtain it
from ExtensionAscanRulesBeta.
Update changes in ZapAddOn.xml file.
Fix zaproxy/zaproxy#1972 - Resource leak while uninstalling "Active
scanner rules (beta)" add-on