fuzz: properly report the number of payloads from the script in dialogue
Change ScriptStringPayloadGeneratorAdapterUIHandler to initialise and
obtain the number of payloads form the script. Also warn if the script
does not implement the expected interface or if an error occurs while
obtaining the number of payloads.
Change ScriptStringPayloadGeneratorAdapter to reuse the script already
initialised by ScriptStringPayloadGeneratorAdapterUIHandler.
Update changes in ZapAddOn.xml.
Fix zaproxy/zaproxy#1887 - Number of payloads not shown for script
payload generator