fuzz: allow to preview payloads generated or from external sources
Change File, File Fuzzers, Regex and Script payload generator UIs to
allow to preview the payloads that would be generated with the selected
options. The payloads are shown in a text area under the generator
options, regex and script generators require user interaction to show
them (as it's not efficient to generate the payloads in realtime with
regex and with script it requires its validation).
The preview does not allow to see all generated payloads since those
might consume too much memory (more than what is actually available),
the number of payloads shown depends on the type of generator but all
allow to have an idea of what it will be generated or read from external
sources.
Update changes in ZapAddOn.xml.
Fix zaproxy/zaproxy#1896 - Enhancement: User Story - I can view payload
file contents while selecting payload files