fuzz: change CompositeIterator to reset the iterators on reset call
Change CompositePayloadGenerator$CompositeIterator to reset the existing
iterators when reset() is called instead of creating new iterators, it
might be more expensive to (re)create the iterators than just reset the
existing ones (for example, creation of a ScriptPayloadGeneratorIterator
might require instantiate again the script, instead of just calling its
reset method). The change also ensures that the iterators are always
closed.