fuzz: allow to modify the selected payload processor script
Change ScriptStringPayloadProcessorAdapterUI to allow to modify the
selected payload processor script otherwise it requires to remove and
add the payload processor script (and move it to correct position).
Change ScriptStringPayloadProcessorAdapterUIPanel to correctly select
the payload processor script when showing the modify dialogue (by
selecting using a ScriptUIEntry instead of ScriptWrapper) and to not
deselect the scripts when clearing (so that there's always a script
selected when adding).
Update changes in ZapAddOn.xml.