fuzz: allow to save (to file) String, Regex, File and Script payloads
Change String, Regex, File and Script payload generators to allow to
save the generated payloads to a file, that's available for selection in
the custom files of File Fuzzers. The generated payloads can be saved
using a button shown in the panel of the generators, which allows to
choose the name of the file (the location is always the same, the one
used to save the custom file fuzzers).
Update changes in ZapAddOn.xml.
Fix zaproxy/zaproxy#1932 - Enhancement: Save user-created payloads for
reuse