fuzz: fix issues and improve support of regex payload generator
Replace the implementation used to generate payloads from regular
expressions with library Generex, it allows to iterate over the
possible generated strings, supports repetitions, returns the expected
strings when a max length is specified (and it also fixes the infinite
loop issue).
Remove class RegExStringGenerator, replaced by the library.
Change the class RegexPayloadGenerator to make use of the library and
drop the use of max length, now that can be specified using repetitions.
Change RegexPayloadGeneratorUI and RegexPayloadGeneratorUIPanel to also
drop the use of max length and to cache the RegexPayloadGenerator as it
contains the number of strings that can be generated (which can be time
consuming to calculate). To prevent infinite loops with infinite regular
expressions (and non-infinite if generated millions of strings) there's
a hardcoded limit of 10M for the calculation of the number of strings
that can be generated and a maximum of 5 seconds to validate/process a
regular expression.
Add more validations to RegexPayloadGeneratorUIPanel, to inform the user
of possible issues of the provided regular expression (for example,
it's infinite and a limit was not specified, or it's not syntactically
valid, or valid but not able to be processed).
Add libraries:
 - Generex - Apache License, Version 2.0 (custom build with required
 enhancements to validate a regex and calculate number of strings with
 iterator);
 - dk.brics.automaton - BSD license (dependency of Generex).
Update classpath, to include the libraries.
Update changes in ZapAddOn.xml.
Fix zaproxy/zaproxy#1884 - Generating regex fuzz payloads takes forever
Fix zaproxy/zaproxy#1885 - Support repetitions in regex payload
generator