fuzz: allow to modify the payloads of file and file fuzzers generators
Change File and File Fuzzers payload generators to allow to modify the
payloads read/generated from the file(s). The payloads can be changed
in the modify dialogue of the corresponding generator, through a text
area. To not change the original file(s) nor keep the payloads in memory
longer than needed it's created a temporary file, used to save and load
the payloads as needed. The edition from within ZAP is limited to a
maximum number of payloads and size of the file that contains them (thus
limiting the size of the payloads). If the file is outside the limits
it's edited with an external text editor, if available.
Update changes in ZapAddOn.xml.
Fix zaproxy/zaproxy#1897 - Enhancement: User Story - I can view and
modify the contents of the payloads I selected