httpsInfo: allow to dynamically update/uninstall the add-on
Change class (extension) RightClickMenu to declare that it can be
dynamically unloaded and move the management of the SSL/TLS dialogues to
itself to keep track of dialogues open (to be closed when the add-on is
unloaded).
Change class MenuEntry to use the extension to show the dialogues.
Bump version and update changes in ZapAddOn.xml file.