zest: fix NullPointerException when active scanning a sequence script
Change class ZestSequenceRunner to use a custom StructuralSiteNode (not
backed by a HistoryReference) for the nodes to be scanned, to return a
custom name and a URI, preventing the NullPointerException(s) from
happening.
The issue was caused by a change in Target's implementation which
started to use a StructuralSiteNode instead of a plain SiteNode.
As part of the changes the nodes to be scanned and the Target were
changed to use a custom name. The custom name (name of the script plus
" (script)") is used, in the active scan tab and progress dialogue, to
properly indicate what is being scanned, otherwise it would not show
anything (that is, an empty string). The custom URI is not actually
scanned it's just used for comparisons with sibling nodes (which never
happen since there are just one "host" node).
Update changes in ZapAddOn.xml file.
Fix zaproxy/zaproxy#2120 - Zest "Active scan sequence" not working