scripts: force compatibility with older ZAP versions (>= 2.4.0)
Change class CommandPanel to force the use of an older constructor of
RTextScrollPane so that the add-on can still be built for older ZAP
versions (>= 2.4.0). The change is required because the new version of
RSyntaxTextArea library introduced a new constructor which would be used
instead of the old constructor (leading to NoSuchMethodError).
Update version (and remove old changes) in ZapAddOn.xml file.