revisit: unload extension and reset revisit state on session changes
Change class ExtensionRevisit to declare that it can be dynamically
unloaded, allowing to dynamically update/uninstall the add-on, and on
unloading remove the RevisitAPI, close/dispose the "revisit setup"
dialogue and remove icons added to Sites tree. Also, change to reset
the revisit state when creating/loading the session and to declare its
dependencies.
Bump version and update changes in ZapAddOn.xml file.