spiderAjax: change AjaxSpiderParam to clone its configuration
Apply the fix done to core class AbstractParam (that fixes its clone
method by cloning the configuration) so that the AJAX Spider add-on can
be fixed without waiting for a core release.
Bump version and update changes in ZapAddOn.xml file.
Related to zaproxy/zaproxy#2151 - AJAX Spider does not click all
elements set in the options