accessControl: fix NullPointerException in AccessControlStatusPanel
Change AccessControlStatusPanel to return immediately when the user
cancels the "Save" dialogue (for the access control report), otherwise
the not selected (null) file would be used in an error message (which
assumed that the report was generated but failed) leading to the
exception. Also, log the exception if it's not possible to create the
XML document during generation of the report (instead of print it to
standard output).
Bump version and update changes in ZapAddOn.xml file.