ascanrules: reduce log level from INFO to DEBUG (and TRACE to DEBUG)
Change scanners CodeInjectionPlugin, CommandInjectionPlugin and
TestExternalRedirect to log the messages when an alert is found with
DEBUG level instead of INFO, to reduce the number of messages logged by
default (those messages are not really needed for everyday use, the
absence or presence of an alert already indicates that fact, if an error
occurred while raising the alert it would be logged by core code with
higher level, ERROR). For CodeInjectionPlugin also change TRACE log
level messages to DEBUG, the former is not used.