ascanrules: tweak PowerShell control pattern to reduce false positives
Change PowerShell pattern in CommandInjectionPlugin to require the
presence of a whitespace character before "get-help", to reduce false
positives when the attack/input ("get-help") is simply reflected in the
response. Also, change the token "cmdlets" to "cmdlet" to increase the
chance of matching chunks of the output of the command.