wappalyzer: allow to dynamically update/uninstall the add-on
Change class ExtensionWappalyzer to declare that it can be dynamically
unloaded and change class PopupMenuEvidence to remove its sub menus from
the list of the main pop up menu once the menu is dismissed (otherwise
it would prevent the objects of the add-on from being GC'ed after it has
been unloaded).
Bump version and update changes in ZapAddOn.xml file.