sequence: tidy up, dynamic unload, finish internationalisation
Allow to dynamically update/uninstall the add-on (just missing the
declaration that it can be unloaded and the removal of the script type).
Declare its (hard) dependencies (ExtensionScript).
Finish internationalisation (some strings that were not yet
internationalised and remove one that was but was not being shown in the
GUI).
Remove hidden column by implementing a custom TableModel, to show the
scripts available in the "Sequence" panel.
Verify and warn if the scripts do not implement SequenceScript
interface.
Initialise view components only when ZAP has a view.
Remove unnecessary type arguments, casts and null checks.
Bump version and update changes in ZapAddOn.xml file.