AlertFilters Minor UI Improvements2
Builds on Pull Request 286.

In ContextAlertFilterPanel:
The Enabled column has a max width based on the optimal width 
for the header content, (can be made smaller, but not bigger).
This means that the dialog can be resized, and the main
content columns will resize, while Enabled stays static based
on the optimal length for the i18n heading being displayed.