Do not allow to dynamically unload Access Control
Change the method ExtensionAcessControl.canUnload() to return false,
the extension is not completely unloaded by default (Context factories
are not unloaded, requires core changes), which might lead to issues,
for example, after uninstallation it would no longer be possible to
create new contexts as it leads to a MissingResourceException.
Update changes in ZapAddOn.xml file.