websocket: correct name of error resource message
Fix the name of the resource message used to indicate that the "fuzzers"
directory is not writable, which was leading to an exception:
ERROR org.zaproxy.zap.ZAP$UncaughtExceptionLogger  - Exception in thread
 "AWT-EventQueue-0"
java.util.MissingResourceException: Can't find resource for bundle
 java.util.PropertyResourceBundle, key fuzz.options.add.dirperms.error
 at java.util.ResourceBundle.getObject(ResourceBundle.java:450)
 at java.util.ResourceBundle.getString(ResourceBundle.java:407)
 at org.zaproxy.zap.extension.fuzz.FuzzOptionsPanel$3.actionPerformed
 (FuzzOptionsPanel.java:330)

The name is "fuzz.options.add.file.dirperms.error" not the above.

Add change to ZapAddOn.xml file.