websocket: adjust log levels, from info to debug
Change to log trivial actions (like establish a WebSocket channel,
receive/insert a message or sending a custom message) as DEBUG instead
of INFO, to not flood the logs.
For example, the log file would be flooded with messages like:
  INFO TableWebSocket  - insert channel: example.com:80 (#1)
  INFO WebSocketProxyV13  - got text frame payload
  INFO TableWebSocket - insert message: #1.1
  INFO WebSocketProxyV13  - got text frame payload
  INFO TableWebSocket  - insert message: #1.2
  INFO WebSocketProxyV13  - got text frame payload
  INFO TableWebSocket  - insert message: #1.3
  INFO WebSocketProxyV13  - got text frame payload
  INFO TableWebSocket  - insert message: #1.4
  INFO WebSocketProxy  - send custom message
  INFO TableWebSocket  - insert message: #1.5

(Thread name and package were stripped from the previous excerpt for
clarity.)

Update changes in ZapAddOn.xml file.