Update FuzzDB add-on with latest files
Update the files from:
 fuzzdb-project/fuzzdb@f1ef0a17bd8d5d4120aeb380f3bd074210781067

Change class FuzzdbFiles to remove the version number from the directory
used for FuzzDB files, the files are no longer released that way.
Changes in ZapAddOn.xml file:
 - Bump version;
 - Normalise the name of the add-on (FuzzDB instead of Fuzzdb);
 - Update description to no longer use the version number;
 - Update changes;
 - Update the URL of the project (moved from Google Code to GitHub);
 - Update "files" entries.

Change Eclipse's classpath to exclude the "files" directories of the
add-ons, prevents compiler errors because of some of the files included
by FuzzDB (the files under "files" directory are not intended to be
compiled).
Add "deploy-fuzzdb" target to build.xml file, to deploy FuzzDB add-on.