coreLang: fix installation issue
Change class ExtensionCoreLanguages to use the file path defined in the
manifest (ZapAddOn.xml) instead of hard coding it, to make sure the file
path/name is always kept in sync and prevent installation issues because
of missing (non existing) file.
Bump version and update changes in ZapAddOn.xml file.