sqliplugin: check all DB technologies available
Change SQLInjectionPlugin scanner to check all DB technologies available
when evaluating if the scanner should be run or not, allowing to run it
even if it's only selected a specific DB technology (not just when it's
selected the top DB technology).
Bump version and update changes in ZapAddOn.xml file.