fuzz: skip fuzzers dir enumeration if not present
Change ExtensionFuzz to not attempt to enumerate the fuzzers directory
if the directory does not exist. Also, change FuzzOptionsPanel to create
the directory if it does not exist when importing/adding a fuzzer file.
Update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#2387 - Fuzzer Extension exception when removing PnH
Extension