fuzz: allow to search fuzzer file names
Change class FuzzerPayloadGeneratorUIPanel to allow to search the names
of the fuzzer files when selecting them, the change adds a find bar to
the top of the tree that shows the available files. The find bar/panel
contains a text field, for the search term, and two buttons, to navigate
the matching tree nodes (next/previous).

Fix zaproxy/zaproxy#2227 - Fuzzer: option to search for payload names