scripts: discard undoable edits no longer needed
Change class CommandPanel to discard the undoable edits of the text area
when no longer needed, that is, when the text area is cleared and when a
new script is set, preventing the edits from accumulating and exhausting
the whole memory (it also prevents undoing the changes to previous
scripts or templates).
Bump version and update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#2675 - Memory exhaustion with Script Console and big
scripts