Add URL Import API
Add ImportUrlsAPI which accepts a simple parameter filePath. Expecting a
ZAP URL export or plain text file containing a single URL per line.

Adjust ExtensionImportUrls:
- hook, now includes registration of the API implementor.
- importUrlFile, modified slightly. It had been assuming that the method
on every line of ZAP URL exports was GET. Now instead of assuming ZAP
formatted exports start with GET it checks if the line starts with http.
If not it assumes everything up to the first tab is the method.
- Update help file to include mentioning the web API endpoint.

Note: The import function still assumes that all URLs are accessible via 
GET. This should largely be true, and it is doubtful that we would find
much benefit in trying the original (exported) method, because for
example on a POST we wouldn't have the necessary parameters (same for a
PUT, or some sort of webdav method).

Fixes zaproxy/zaproxy#1643