websocket: fix overlapping of text location
Change TextWebSocketMessageLocation to check if the new text location
is (or not) an insertion point, when the existing text location has the
same start and different end. If it's not an insertion point it's
overlapping the existing location.
Update changes in ZapAddOn.xml file.

Related to zaproxy/zaproxy#2792 - Able to overlap fuzz (HTTP) locations