zest: show always the expected request URL
Change ZestRequestDialog to show the URL with tokens if defined, if not,
show "plain" URL (if defined, otherwise fallback to empty URL). Also,
reset the "plain" URL when setting the URL with tokens and reset the URL
with tokens when setting the "plain" URL, which ensures that it's shown
and used the correct URL.
Change ZestZapRunner to throw an exception if the request-uri is not set
(e.g. the URL is not well-formed).
Update changes in ZapAddOn.xml file (remove existing code change
mention, it's no longer needed or of interest).
Fix zaproxy/zaproxy#2854 - Cannot use variable in Zest Script as part of
POST URL