ascanrulesBeta: fix attack URL in PHP source scans
Change scanner SourceCodeDisclosureCVE20121823 to create the attack URL
using the original URL components (authority and path) escaped, as
that's the expected form, otherwise it would lead to an exception:
 "escaped absolute path not valid"
if there were characters that required escaping.

Add tests to assert the expected behaviour of the scanner.
Update changes in ZapAddOn.xml file.
Change build.xml file to include ZAP add-ons when compiling the tests
(to include required dependencies).