ascanrules: improve error handling in scanners
Change scanners BufferOverflow, FormatString and TestCrossSiteScriptV2
to catch URIException when sending a message as that can be thrown if
the URI is not correct (non HTTP/S scheme or invalid encoding).
Also, change TestCrossSiteScriptV2 to not attempt to use the message if
it was not successfully sent (e.g. because of the above exception).
Bump version and update changes in ZapAddOn.xml file.