fuzz: fix exception when adding file fuzzers
Change FuzzerPayloadGeneratorUIHandler to check that the tree leaf nodes
being added are of the expected type (the top level nodes might be empty
thus being a leaf and treated as source of file fuzzers, which aren't).
The exception would happen, for example, when adding file fuzzers with
empty "Custom Fuzzers" node selected.
Update changes in ZapAddOn.xml file.