zest: add reqs to Sequence scripts when recording
Change ExtensionZest to also add the requests to Sequence scripts,
moreover add the requests to any script that has the "append"
capability, that is, the ones that can be recorded.
Change ZestRecordOnOffPopupMenu to be enabled to all the scripts that
can be recorded.
Update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#3044 - Recording of Zest Sequence scripts not
working