ascanrulesbeta: fix src disclosure false positive
Change SourceCodeDisclosureWEBINF to not raise an alert if the
successful response does not decompile correctly to a Java class.
Add tests to assert (some of) the expected behaviour of the scanner.
Bump version and update changes in ZapAddOn.xml file.

Also, change how path segment is checked in test server to allow the
tests to precisely define what the path is.
Update existing unit tests to match the new behaviour of the test
server.
Fix typo in JavaDoc of helper test method.

Fix zaproxy/zaproxy#3048 - False positive on source disclosure when
target answers invalid url with 200 OK