Add some tests for CommandInjectionPlugin
Add tests for CommandInjectionPlugin to check that the value used in
time-based attacks is the expected.
Change class ActiveScannerTest to be generic so the instance variable
"rule" has the type of the scanner being tested (to allow to use its
methods without casts).
Update existing unit test, TestCrossSiteScriptV2UnitTest, to set the
type argument for ActiveScannerTest (and remove unnecessary casts).