jxbrowser: provide WebDrivers
Change JxBrowser add-ons to provide WebDrivers, so that it can be used
by other add-ons (e.g. AJAX Spider).
Change some classes to allow to create tabs with Browsers, to have just
one window/dialogue for all browsers required by automated tools.

The JxBrowserProvider implementation is duplicated as core does not
support the sharing of the class in other non-dependent add-ons. The
duplication will be removed once core supports that.
The JxBrowser add-ons do not depend (directly) on the Selenium add-on,
allowing the JxBrowser to be used without Selenium add-on.