selenium: document the IDs of the browsers
Update help page to mention the IDs of the browsers, needed for choosing
a browser through the command line or the ZAP API. Also, mention that
other add-ons can add browsers.
The help page was reorganised to have the name, ID and requirements of
the browsers in the same place (table).
Bump version and update changes in ZapAddOn.xml file.