scripts: inform about scripts with wrong charset
Change ScriptsListPanel to inform the user when the script contains
invalid character sequences and that it needs to be in UTF-8, instead
of showing the generic error.
Bump versions (add-on/ZAP) and update changes in ZapAddOn.xml file.

Related to zaproxy/zaproxy#3377 - ZAP can no longer load non UTF8
scripts