openapi: fallback to scheme of request URI
Change OpenApiSpider and ExtensionOpenApi to use the scheme of the
request URI, used to obtain the spec, when the scheme is not specified
in the spec.
Bump version and update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#3433 - OpenAPI: Parsing of JSON downloaded from a
URL fails if no schemes element is present