zest: select correct script when editing actions
Change ZestActionDialog to select the correct script of the action by
using the name of the script instead of the file path. If the script no
longer exists (that is, added to ZAP) it's selected an existing one.
Update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#3489 - Zest script execution UI does not show the
correct configuration