Replace use of regex in TestPathTraversal dir test
Change TestPathTraversal to simply search for the directory names
instead of using a regular expression, while with the regular expression
the evidence included all the directory names it was not efficient with
bigger responses (new behaviour returns one of the directory names).
For example, with a 76k chars response the regex took 153s to complete
but less than 1s with the simple search.
Add tests to assert the expected behaviour when testing for directory
listings.

Fix zaproxy/zaproxy#3468 - PathTraversal rule performance issue