alertFilters: dynamically unload the add-on
Change ExtensionAlertFilters to add the components through the provided
hook (instead of adding them directly to the core components, so that
they are automatically unloaded when the add-on is uninstalled) and
declare that the extension can be unloaded.
Bump minimum ZAP version and version of the add-on and update changes in
ZapAddOn.xml file.