domxss: use Selenium extension and add skip reason
Change ExtensionDomXSS to depend on ExtensionSelenium and to dynamically
add/remove the active scanner (TestDomXSS), to ensure the extension is
always available to the scanner.
Change TestDomXSS to use ExtensionSelenium to obtain the WebDriver,
which uses the latest/correct WebDriver implementation to connect to
Firefox (geckodriver/marionette instead of xpi add-on). Also, set the
reason why the scanner was skipped when failed to use the browser.
Update changes, minimum Java/ZAP version and Selenium add-on version in
ZapAddOn.xml file.

Fix zaproxy/zaproxy#3396 - Enhance DOM XSS scanner addon to use Selenium
addon
Related to zaproxy/zaproxy#3509 - Update Selenium library to version 3.x