quickstart: update code to ZAP 2.7.0 (#1209)
Change ExtensionQuickStartLaunch to use the API core method now that
the add-on is targeting 2.7.0. Set the name of the threads to be easier
to identify them.
Remove code no longer needed in QuickStartLaunchAPI per previous API
change.
Change QuickStartPanel to set the state that indicates the panel is
shown by default (instead of overriding the method).
Bump version and update changes in ZapAddOn.xml file.