HttpsInfo UI Tweak (#1486)
This minor change changes the tab close functionality so that tabs are
actually removed not just hidden. As this is related to the previous PR
I have not included any further "changes" in the manifest.

I decided to go ahead with this changes as the 5.1.0beta integration has
been stalled due to some performance issues that I've had to go back to
the author with and am now waiting on a new build.