tokengen: react to session/mode changes
Change ExtensionTokenGen to stop the test, close the dialogues, and
reset the status panel when the session changes. Also, react to mode
changes, stop the test when changed to safe or protected if the message
is not in scope.
Change TokenPanel to also set the main panel (instructions) and reset
the state of the save/pause buttons when it is reset.
Change GenerateTokensDialog to prevent the token generation when in safe
mode or protected if the message is not in scope.
Update help to mention the mode restrictions.
Update changes in ZapAddOn.xml file.